Issue #2557: made tests only cover one subject
